requirements.txt
Flask==3.0.0
requests==2.31.0
pytest==7.4.3
black==24.10.0
flake8==7.0.0
pylint==3.0.3
pre-commit==3.6.0
python-dotenv==1.0.0